:white_check_mark:

+++ integration-tests/src/test/java/cn/taketoday/web/client/config/ClientHttpRequestFactoriesHttpComponentsTests.java
@@ -1,8 +1,5 @@
 /*
- * Original Author -> Harry Yang (taketoday@foxmail.com) https://taketoday.cn
- * Copyright © Harry Yang & 2017 - 2023 All Rights Reserved.
- *
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER
+ * Copyright 2017 - 2023 the original author or authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -43,7 +40,7 @@ class ClientHttpRequestFactoriesHttpComponentsTests
 
   @Override
   protected long connectTimeout(HttpComponentsClientHttpRequestFactory requestFactory) {
-    return (int) ReflectionTestUtils.getField(requestFactory, "connectTimeout");
+    return (long) ReflectionTestUtils.getField(requestFactory, "connectTimeout");
   }
 
   @Override
